requirements.txt
streamlit
pandas
matplotlib
numpy
seaborn
openpyxl
PyGithub
